@@ -20,11 +20,12 @@
   "homepage": "https://github.com/newebio/neweb-core#readme",
   "jest": {},
   "devDependencies": {
-    "@types/jest": "^22.2.3",
+    "@types/jest": "^23.1.1",
     "coveralls": "^3.0.1",
-    "jest": "^22.4.3",
+    "jest": "^23.2.0",
+    "sleep-es6": "^0.0.2",
     "tslint": "^5.10.0",
-    "typescript": "^2.8.3"
+    "typescript": "^2.9.2"
   },
   "dependencies": {
     "@types/debug": "^0.0.30",
