@@ -1,21 +1,46 @@
 import debug = require("debug");
-import { BehaviorSubject, Subscription } from "rxjs";
-import { skip, take } from "rxjs/operators";
+import { Subscription } from "rxjs";
 import {
-    IController, IControllerMessageParams,
-    IControllersFactory, IPage, IRoute, IRoutePage, IRouter,
-    IRoutersFactory, ISeance, ISeanceClient, ISeanceInitializeParams,
+    IController,
+    IControllerMessageParams,
+    IControllersFactory,
+    IPage,
+    IRoute,
+    IRoutePage,
+    IRouter,
+    IRoutersFactory,
+    ISeance,
+    ISeanceClient,
+    ISeanceInitializeParams,
+    IPageCreator,
 } from "./typings";
-import { fillPage, routeToPage } from "./util/page";
+import { fillPage } from "./util/page";
 import PageComparator from "./util/PageComparator";
+import PageCreator from "./util/PageCreator";
 
 export interface ISeanceConfig {
-    RoutersFactory: IRoutersFactory;
+    url: string;
     ControllersFactory: IControllersFactory;
+    router: IRouter;
+    PageCreator: IPageCreator;
 }
 export class Seance implements ISeance {
-    public page$: BehaviorSubject<IPage>;
-    public url$: BehaviorSubject<string>;
+    public static async create(
+        params: ISeanceInitializeParams & {
+            RoutersFactory: IRoutersFactory;
+            ControllersFactory: IControllersFactory;
+        },
+    ) {
+        debug("seance")("create new seance", params);
+        const router = await params.RoutersFactory.createRouter();
+        const pageCreator = new PageCreator();
+        return new Seance({
+            url: params.url,
+            router,
+            ControllersFactory: params.ControllersFactory,
+            PageCreator: pageCreator,
+        });
+    }
     protected client?: ISeanceClient;
     protected router: IRouter;
     protected controllers: {
@@ -24,67 +49,81 @@ export class Seance implements ISeance {
             subscriptions: Subscription[];
         };
     } = {};
-    constructor(protected config: ISeanceConfig) { }
-    public async initialize(params: ISeanceInitializeParams) {
-        debug("seance")("initialize", params);
-        this.url$ = new BehaviorSubject(params.url);
-        this.router = this.config.RoutersFactory.createRouter({ url$: this.url$ });
-        const route = await this.router.route$.pipe(take(1)).toPromise();
-        if (route.type !== "page") {
-            throw new Error("");
-        }
-        // Create non-filled page from routePage
-        const page = await routeToPage(route.page);
-        const controllers = await fillPage(this.config.ControllersFactory, page);
-        controllers.map(({ id, controller }) => {
-            const subscriptions: Subscription[] = [];
-            subscriptions.push(
-                controller.onMessage.subscribe((message) => this.onControllerMessage({
-                    id,
-                    message,
-                })),
-            );
-            this.controllers[id] = { controller, subscriptions };
-        });
-        this.page$ = new BehaviorSubject(page);
-        this.router.route$.pipe(skip(1)).subscribe((r) => this.onNewRoute(r));
+    protected page?: IPage;
+    constructor(protected config: ISeanceConfig) {
+        this.router = config.router;
+        this.router.onChangeRoute.subscribe(this.onNewRoute);
+        this.router.emitNewUrl.next(config.url);
     }
     public connect(client: ISeanceClient) {
         this.client = client;
         this.client.onNavigate.subscribe((params) => this.navigate(params.url));
         this.client.emitControllerMessage.subscribe(this.emitControllerMessage);
-        this.page$.subscribe((page) => client.emitNewPage.next({ page }));
+        if (this.page) {
+            this.client.emitNewPage.next({ page: this.page });
+        }
     }
-    protected async onNewRoute(route: IRoute) {
+    public async dispose() {
+        await this.router.dispose();
+    }
+    protected onNewRoute = async (route: IRoute) => {
         debug("seance")("new route", route);
         if (route.type === "redirect") {
-            this.navigate(route.url);
+            this.page = undefined;
+            await this.navigate(route.url);
+            return;
+        }
+        if (route.type === "notFound") {
+            // TODO
+            this.page = undefined;
             return;
         }
         if (route.type === "page") {
-            this.page$.next(await this.replacePage(this.page$.getValue(), route.page));
-            return;
+            await this.onNewRoutePage(route.page);
+        }
+    };
+    protected async onNewRoutePage(routePage: IRoutePage) {
+        // Create non-filled page from routePage
+        const page = this.page ? await this.replacePage(this.page, routePage) : await this.createPage(routePage);
+        if (this.client) {
+            this.client.emitNewPage.next({ page });
         }
     }
-    protected navigate = (url: string) => {
-        this.url$.next(url);
+    protected async createPage(routePage: IRoutePage) {
+        const page = await this.config.PageCreator.createPage(routePage);
+        const controllers = await fillPage(this.config.ControllersFactory, page);
+        controllers.map(({ id, controller }) => {
+            const subscriptions: Subscription[] = [];
+            subscriptions.push(
+                controller.onMessage.subscribe((message) =>
+                    this.onControllerMessage({
+                        id,
+                        message,
+                    }),
+                ),
+            );
+            this.controllers[id] = { controller, subscriptions };
+        });
+        return page;
     }
+    protected navigate = (url: string) => {
+        this.router.emitNewUrl.next(url);
+    };
     protected onControllerMessage = (params: IControllerMessageParams) => {
         debug("seance")("onControllerMessage", params);
         if (this.client) {
             this.client.onControllerMessage.next(params);
         }
-    }
+    };
     protected emitControllerMessage = (params: IControllerMessageParams) => {
         debug("seance")("emitControllerMessage", params);
         const controllerItem = this.controllers[params.id];
         if (controllerItem) {
             controllerItem.controller.postMessage.next(params.message);
         }
-    }
-
+    };
     protected async replacePage(oldPage: IPage, routePage: IRoutePage): Promise<IPage> {
-        const newPage = await routeToPage(routePage);
+        const newPage = await this.config.PageCreator.createPage(routePage);
         const pageComparator = new PageComparator();
         const info = pageComparator.getCompareInfo(oldPage, newPage);
         // TODO WAIT?
@@ -95,24 +134,32 @@ export class Seance implements ISeance {
                 return controller.onChangeParams.next(frame.params);
             }
         });
-        await Promise.all<any>(info.newFrames.map(async (frame) => {
-            const controller = await this.config.ControllersFactory.create(frame.frameName);
-            frame.data = await controller.init();
-            const id = frame.frameId;
-            const subscriptions: Subscription[] = [];
-            subscriptions.push(controller.onMessage.subscribe((message) => this.onControllerMessage({
-                id,
-                message,
-            })));
-            this.controllers[id] = { controller, subscriptions };
-        }).concat(changeParamsPromises));
+        await Promise.all<any>(
+            info.newFrames
+                .map(async (frame) => {
+                    const controller = await this.config.ControllersFactory.create(frame.frameName);
+                    frame.data = await controller.init();
+                    const id = frame.frameId;
+                    const subscriptions: Subscription[] = [];
+                    subscriptions.push(
+                        controller.onMessage.subscribe((message) =>
+                            this.onControllerMessage({
+                                id,
+                                message,
+                            }),
+                        ),
+                    );
+                    this.controllers[id] = { controller, subscriptions };
+                })
+                .concat(changeParamsPromises),
+        );
         return info.page;
     }
     protected async removeController(id: string) {
         const controllerItem = this.controllers[id];
         if (controllerItem) {
             delete this.controllers[id];
-            if (typeof (controllerItem.controller.dispose) === "function") {
+            if (typeof controllerItem.controller.dispose === "function") {
                 await controllerItem.controller.dispose();
             }
             controllerItem.subscriptions.map((s) => s.unsubscribe());
