@@ -11,6 +11,54 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const querystring = require("querystring");
 const rxjs_1 = require("rxjs");
 const url_1 = require("url");
+class ClassicRouter {
+    constructor() {
+        this.emitNewUrl = new rxjs_1.Subject();
+        this.onChangeRoute = new rxjs_1.Subject();
+        this.routes = [];
+        this.onInit();
+        this.urlSubscription = this.emitNewUrl.subscribe((url) => {
+            this.navigate({
+                request: {
+                    url,
+                },
+            });
+        });
+    }
+    addRoute(route) {
+        this.routes.push(route);
+    }
+    navigate(params) {
+        return __awaiter(this, void 0, void 0, function* () {
+            this.currentRequest = params.request;
+            for (const routeHandler of this.routes) {
+                const route = routeHandler(params.request, {});
+                if (route) {
+                    this.emitRoute(route);
+                    return;
+                }
+            }
+            this.emitRoute({
+                type: "notFound",
+                text: "Unknown request " + params.request.url,
+            });
+        });
+    }
+    dispose() {
+        return __awaiter(this, void 0, void 0, function* () {
+            this.urlSubscription.unsubscribe();
+        });
+    }
+    emitRoute(route) {
+        return __awaiter(this, void 0, void 0, function* () {
+            this.onChangeRoute.next(route);
+        });
+    }
+    onInit() {
+        //
+    }
+}
+exports.ClassicRouter = ClassicRouter;
 function MatchedRoute(opts, next) {
     return (request, context) => {
         const paths = opts.path.split("/");
@@ -18,14 +66,14 @@ function MatchedRoute(opts, next) {
         const paramsNames = [];
         for (const pathO of paths) {
             if (pathO.indexOf(":") === 0) {
-                regexpArr.push("([^\/]+)");
+                regexpArr.push("([^/]+)");
                 paramsNames.push(pathO.substr(1));
             }
             else {
                 regexpArr.push(pathO);
             }
         }
-        const regexp = new RegExp("^" + regexpArr.join("\\\/") + "$", "");
+        const regexp = new RegExp("^" + regexpArr.join("\\/") + "$", "");
         const url = url_1.parse(request.url);
         const pathname = url.pathname || "/";
         const match = pathname.match(regexp);
@@ -103,56 +151,4 @@ function RouteWithRedirectOn(params, next) {
     };
 }
 exports.RouteWithRedirectOn = RouteWithRedirectOn;
-class ClassicRouter {
-    constructor(config) {
-        this.config = config;
-        this.routes = [];
-        this.onInit();
-        this.urlSubscription = this.config.url$.subscribe((url) => {
-            this.navigate({
-                request: {
-                    url,
-                },
-            });
-        });
-    }
-    onInit() {
-        //
-    }
-    addRoute(route) {
-        this.routes.push(route);
-    }
-    navigate(params) {
-        return __awaiter(this, void 0, void 0, function* () {
-            this.currentRequest = params.request;
-            for (const routeHandler of this.routes) {
-                const route = routeHandler(params.request, {});
-                if (route) {
-                    this.emitRoute(route);
-                    return;
-                }
-            }
-            this.emitRoute({
-                type: "notFound",
-                text: "Unknown request " + params.request.url,
-            });
-        });
-    }
-    dispose() {
-        return __awaiter(this, void 0, void 0, function* () {
-            this.urlSubscription.unsubscribe();
-        });
-    }
-    emitRoute(route) {
-        return __awaiter(this, void 0, void 0, function* () {
-            if (!this.route$) {
-                this.route$ = new rxjs_1.BehaviorSubject(route);
-            }
-            else {
-                this.route$.next(route);
-            }
-        });
-    }
-}
-exports.ClassicRouter = ClassicRouter;
 exports.default = ClassicRouter;
