@@ -5,14 +5,14 @@ const operators_1 = require("rxjs/operators");
 class Client {
     constructor(config) {
         this.config = config;
-        this.onControllerMessage$ = new rxjs_1.Subject();
-        this.onNewPage$ = new rxjs_1.Subject();
+        this.onControllerMessage = new rxjs_1.Subject();
+        this.onNewPage = new rxjs_1.Subject();
         // statuses
-        this.networkStatus$ = new rxjs_1.BehaviorSubject("connecting");
-        this.navigateStatus$ = new rxjs_1.BehaviorSubject("navigating");
+        this.onChangeNetworkStatus = new rxjs_1.BehaviorSubject("connecting");
+        this.onChangeNavigateStatus = new rxjs_1.BehaviorSubject("navigating");
         this.connect = () => {
-            this.networkStatus$.next("connected");
-            this.config.transport.outputMessage$.next({
+            this.onChangeNetworkStatus.next("connected");
+            this.config.transport.outputMessage.next({
                 type: "initialize",
                 body: {
                     seanceId: this.seanceId,
@@ -21,43 +21,40 @@ class Client {
             });
         };
         this.disconnect = () => {
-            this.networkStatus$.next("disconnected");
+            this.onChangeNetworkStatus.next("disconnected");
         };
         this.connecting = () => {
-            this.networkStatus$.next("connecting");
+            this.onChangeNetworkStatus.next("connecting");
         };
         this.url = this.config.url;
         // controller actions
-        this.emitControllerMessage$ = new rxjs_1.Subject();
-        this.emitControllerMessage$.subscribe((body) => config.transport.outputMessage$.next({
+        this.emitControllerMessage = new rxjs_1.Subject();
+        this.emitControllerMessage.subscribe((body) => config.transport.outputMessage.next({
             type: "controller-message",
             body,
         }));
         // navigating
-        this.emitNavigate$ = new rxjs_1.Subject();
-        this.emitNavigate$.subscribe((body) => {
-            this.navigateStatus$.next("navigating");
-            config.transport.outputMessage$.next({
+        this.emitNavigate = new rxjs_1.Subject();
+        this.emitNavigate.subscribe((body) => {
+            this.onChangeNavigateStatus.next("navigating");
+            config.transport.outputMessage.next({
                 type: "navigate",
                 body,
             });
         });
-        config.transport.onConnect$.subscribe(this.connect);
-        config.transport.onDisconnect$.subscribe(this.disconnect);
-        config.transport.onConnecting$.subscribe(this.connecting);
-        config.transport.inputMessage$
+        config.transport.onConnect.subscribe(this.connect);
+        config.transport.onDisconnect.subscribe(this.disconnect);
+        config.transport.onConnecting.subscribe(this.connecting);
+        config.transport.inputMessage
             .pipe(operators_1.filter((message) => message.type === "new-page"))
             .subscribe((message) => this.emitNewPage(message.body.page));
-        config.transport.inputMessage$
+        config.transport.inputMessage
             .pipe(operators_1.filter((message) => message.type === "controller-message"))
-            .subscribe((message) => this.onControllerMessage(message.body));
+            .subscribe((message) => this.onControllerMessage.next(message.body));
     }
     emitNewPage(page) {
-        this.navigateStatus$.next("navigating");
-        this.onNewPage$.next(page);
-    }
-    onControllerMessage(params) {
-        this.onControllerMessage$.next(params);
+        this.onChangeNavigateStatus.next("navigating");
+        this.onNewPage.next(page);
     }
 }
 exports.Client = Client;
