@@ -9,16 +9,31 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const debug = require("debug");
-const rxjs_1 = require("rxjs");
-const operators_1 = require("rxjs/operators");
 const page_1 = require("./util/page");
 const PageComparator_1 = require("./util/PageComparator");
+const PageCreator_1 = require("./util/PageCreator");
 class Seance {
     constructor(config) {
         this.config = config;
         this.controllers = {};
+        this.onNewRoute = (route) => __awaiter(this, void 0, void 0, function* () {
+            debug("seance")("new route", route);
+            if (route.type === "redirect") {
+                this.page = undefined;
+                yield this.navigate(route.url);
+                return;
+            }
+            if (route.type === "notFound") {
+                // TODO
+                this.page = undefined;
+                return;
+            }
+            if (route.type === "page") {
+                yield this.onNewRoutePage(route.page);
+            }
+        });
         this.navigate = (url) => {
-            this.url$.next(url);
+            this.router.emitNewUrl.next(url);
         };
         this.onControllerMessage = (params) => {
             debug("seance")("onControllerMessage", params);
@@ -33,18 +48,48 @@ class Seance {
                 controllerItem.controller.postMessage.next(params.message);
             }
         };
+        this.router = config.router;
+        this.router.onChangeRoute.subscribe(this.onNewRoute);
+        this.router.emitNewUrl.next(config.url);
     }
-    initialize(params) {
+    static create(params) {
         return __awaiter(this, void 0, void 0, function* () {
-            debug("seance")("initialize", params);
-            this.url$ = new rxjs_1.BehaviorSubject(params.url);
-            this.router = this.config.RoutersFactory.createRouter({ url$: this.url$ });
-            const route = yield this.router.route$.pipe(operators_1.take(1)).toPromise();
-            if (route.type !== "page") {
-                throw new Error("");
-            }
+            debug("seance")("create new seance", params);
+            const router = yield params.RoutersFactory.createRouter();
+            const pageCreator = new PageCreator_1.default();
+            return new Seance({
+                url: params.url,
+                router,
+                ControllersFactory: params.ControllersFactory,
+                PageCreator: pageCreator,
+            });
+        });
+    }
+    connect(client) {
+        this.client = client;
+        this.client.onNavigate.subscribe((params) => this.navigate(params.url));
+        this.client.emitControllerMessage.subscribe(this.emitControllerMessage);
+        if (this.page) {
+            this.client.emitNewPage.next({ page: this.page });
+        }
+    }
+    dispose() {
+        return __awaiter(this, void 0, void 0, function* () {
+            yield this.router.dispose();
+        });
+    }
+    onNewRoutePage(routePage) {
+        return __awaiter(this, void 0, void 0, function* () {
             // Create non-filled page from routePage
-            const page = yield page_1.routeToPage(route.page);
+            const page = this.page ? yield this.replacePage(this.page, routePage) : yield this.createPage(routePage);
+            if (this.client) {
+                this.client.emitNewPage.next({ page });
+            }
+        });
+    }
+    createPage(routePage) {
+        return __awaiter(this, void 0, void 0, function* () {
+            const page = yield this.config.PageCreator.createPage(routePage);
             const controllers = yield page_1.fillPage(this.config.ControllersFactory, page);
             controllers.map(({ id, controller }) => {
                 const subscriptions = [];
@@ -54,32 +99,12 @@ class Seance {
                 })));
                 this.controllers[id] = { controller, subscriptions };
             });
-            this.page$ = new rxjs_1.BehaviorSubject(page);
-            this.router.route$.pipe(operators_1.skip(1)).subscribe((r) => this.onNewRoute(r));
-        });
-    }
-    connect(client) {
-        this.client = client;
-        this.client.onNavigate.subscribe((params) => this.navigate(params.url));
-        this.client.emitControllerMessage.subscribe(this.emitControllerMessage);
-        this.page$.subscribe((page) => client.emitNewPage.next({ page }));
-    }
-    onNewRoute(route) {
-        return __awaiter(this, void 0, void 0, function* () {
-            debug("seance")("new route", route);
-            if (route.type === "redirect") {
-                this.navigate(route.url);
-                return;
-            }
-            if (route.type === "page") {
-                this.page$.next(yield this.replacePage(this.page$.getValue(), route.page));
-                return;
-            }
+            return page;
         });
     }
     replacePage(oldPage, routePage) {
         return __awaiter(this, void 0, void 0, function* () {
-            const newPage = yield page_1.routeToPage(routePage);
+            const newPage = yield this.config.PageCreator.createPage(routePage);
             const pageComparator = new PageComparator_1.default();
             const info = pageComparator.getCompareInfo(oldPage, newPage);
             // TODO WAIT?
@@ -90,7 +115,8 @@ class Seance {
                     return controller.onChangeParams.next(frame.params);
                 }
             }));
-            yield Promise.all(info.newFrames.map((frame) => __awaiter(this, void 0, void 0, function* () {
+            yield Promise.all(info.newFrames
+                .map((frame) => __awaiter(this, void 0, void 0, function* () {
                 const controller = yield this.config.ControllersFactory.create(frame.frameName);
                 frame.data = yield controller.init();
                 const id = frame.frameId;
@@ -100,7 +126,8 @@ class Seance {
                     message,
                 })));
                 this.controllers[id] = { controller, subscriptions };
-            })).concat(changeParamsPromises));
+            }))
+                .concat(changeParamsPromises));
             return info.page;
         });
     }
@@ -109,7 +136,7 @@ class Seance {
             const controllerItem = this.controllers[id];
             if (controllerItem) {
                 delete this.controllers[id];
-                if (typeof (controllerItem.controller.dispose) === "function") {
+                if (typeof controllerItem.controller.dispose === "function") {
                     yield controllerItem.controller.dispose();
                 }
                 controllerItem.subscriptions.map((s) => s.unsubscribe());
