@@ -9,92 +9,49 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const debug = require("debug");
-const rxjs_1 = require("rxjs");
 const operators_1 = require("rxjs/operators");
-const uid = require("uid-safe");
-const Seance_1 = require("./Seance");
+const SeanceClient_1 = require("./SeanceClient");
+const SeancesManager_1 = require("./SeancesManager");
 class Server {
     constructor(config) {
         this.config = config;
-        this.seances = {};
-        this.config.transport.onConnect$.subscribe((transportClient) => {
+        this.subscriptions = [];
+        this.onNewTransportClient = (transportClient) => {
             debug("server")("new client", transportClient);
-            transportClient.inputMessage$
-                .pipe(operators_1.filter((message) => message.type === "initialize"))
-                .subscribe((message) => this.onSeanceRequest({
-                client: this.createSeanceClient(transportClient),
+            const onInitialize = transportClient.inputMessage.pipe(operators_1.filter((message) => message.type === "initialize"));
+            this.subscriptions.push(onInitialize.subscribe((message) => this.onTransportClientInitialize(transportClient, message)));
+        };
+        this.onTransportClientInitialize = (transportClient, message) => __awaiter(this, void 0, void 0, function* () {
+            const client = this.createSeanceClient(transportClient);
+            debug("server")("initialize", message);
+            const seance = yield this.config.seancesManager.resolveSeance({
                 seanceId: message.body.seanceId,
-                extraInfo: transportClient.getExtraInfo(),
                 sessionId: transportClient.getSessionId(),
                 url: message.body.url,
-            }));
+                extraInfo: transportClient.getExtraInfo(),
+            });
+            debug("server")("connect seance");
+            seance.connect(client);
         });
     }
-    onSeanceRequest(params) {
-        return __awaiter(this, void 0, void 0, function* () {
-            debug("server")("initialize", params);
-            const seance = yield this.resolveSeance(params);
-            debug("server")("connect seance");
-            seance.connect(params.client);
+    static create(config) {
+        const seancesManager = new SeancesManager_1.default({
+            ControllersFactory: config.ControllersFactory,
+            RoutersFactory: config.RoutersFactory,
         });
+        return new Server({
+            seancesManager,
+            transport: config.transport,
+        });
+    }
+    start() {
+        this.subscriptions.push(this.config.transport.onConnect.subscribe(this.onNewTransportClient));
+    }
+    stop() {
+        this.subscriptions.map((s) => s.unsubscribe());
     }
     createSeanceClient(transportClient) {
-        const seanceClient = {
-            onControllerMessage: new rxjs_1.Subject(),
-            emitNewPage: new rxjs_1.Subject(),
-            emitControllerMessage: transportClient.inputMessage$
-                .pipe(operators_1.filter((message) => message.type === "controller-message"), operators_1.map((message) => message.body)),
-            onNavigate: transportClient.inputMessage$
-                .pipe(operators_1.filter((message) => message.type === "navigate"), operators_1.map((message) => message.body)),
-        };
-        seanceClient.onControllerMessage.subscribe((body) => {
-            transportClient.outputMessage$.next({
-                type: "controller-message",
-                body,
-            });
-        });
-        seanceClient.emitNewPage.subscribe((body) => transportClient.outputMessage$.next({
-            type: "new-page",
-            body,
-        }));
-        return seanceClient;
-    }
-    resolveSeance(params) {
-        return __awaiter(this, void 0, void 0, function* () {
-            if (params.seanceId) {
-                const seanceItem = this.seances[params.seanceId];
-                if (seanceItem) {
-                    if (seanceItem.sessionId === params.sessionId) {
-                        seanceItem.lastAccessTime = new Date();
-                        return seanceItem.seance;
-                    }
-                }
-            }
-            return this.createSeance(params);
-        });
-    }
-    createSeance(params) {
-        return __awaiter(this, void 0, void 0, function* () {
-            const seanceId = yield this.generateSeanceId();
-            const seance = new Seance_1.default({
-                ControllersFactory: this.config.ControllersFactory,
-                RoutersFactory: this.config.RoutersFactory,
-            });
-            yield seance.initialize(params);
-            const now = new Date();
-            this.seances[seanceId] = {
-                seance,
-                sessionId: params.sessionId,
-                createdAt: now,
-                lastAccessTime: now,
-            };
-            return seance;
-        });
-    }
-    generateSeanceId() {
-        return __awaiter(this, void 0, void 0, function* () {
-            return uid(20);
-        });
+        return new SeanceClient_1.default({ transportClient });
     }
 }
 exports.Server = Server;
