@@ -1,51 +1,36 @@
 "use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
 Object.defineProperty(exports, "__esModule", { value: true });
+const generateFrameId_1 = require("./generateFrameId");
 class PageCreator {
     constructor() {
         this.nonce = 0;
     }
     createPage(routePage) {
-        return __awaiter(this, void 0, void 0, function* () {
-            const frames = this.collectFrames(routePage.rootFrame);
-            const pageFrames = yield Promise.all(frames.allFrames
-                .map((frame) => this.createFrame(frame.frameId, frame.frame, frame.frames)));
-            return {
-                url: routePage.url,
-                frames: pageFrames,
-                rootFrame: frames.frameId,
-                extraInfo: {},
-            };
-        });
+        const frames = this.collectFrames(routePage.rootFrame);
+        const pageFrames = frames.allFrames.map((frame) => this.createFrame(frame.frameId, frame.frame, frame.frames));
+        return {
+            url: routePage.url,
+            frames: pageFrames,
+            rootFrame: frames.frameId,
+            extraInfo: {},
+        };
     }
     createFrame(frameId, routePageFrame, children) {
-        return __awaiter(this, void 0, void 0, function* () {
-            if (!frameId) {
-                frameId = this.generateFrameId();
-            }
-            const frameName = routePageFrame.name;
-            const params = routePageFrame.params;
-            return {
-                frameId,
-                frameName,
-                frames: children,
-                data: {},
-                params,
-            };
-        });
-    }
-    generateFrameId() {
-        return (+new Date()).toString() + Math.round(Math.random() * 10000).toString() + ++this.nonce;
+        if (!frameId) {
+            frameId = generateFrameId_1.generateFrameId(++this.nonce);
+        }
+        const frameName = routePageFrame.name;
+        const params = routePageFrame.params;
+        return {
+            frameId,
+            frameName,
+            frames: children,
+            data: {},
+            params,
+        };
     }
     collectFrames(routePageFrame) {
-        const frameId = this.generateFrameId();
+        const frameId = generateFrameId_1.generateFrameId(++this.nonce);
         let allFrames = [];
         const frames = {};
         Object.keys(routePageFrame.frames).map((framePlace) => {
